+++ CHANGELOG.md
@@ -11,11 +11,12 @@
 - `storeID` added to atom effects interface (#1414)
 - `<RecoilRoot>` will only call `initializeState()` during initial render. (#1372)
 - `RecoilLoadable.all()` and `RecoilLoadable.of()` accept either literal values, async Promises, or Loadables.
+- Selector `getCallback()` callbacks can now mutate, refresh, and transact Recoil state in addition to reading it.
+- Re-renders from Recoil updates now occur 1) earlier, 2) in sync with React updates in the same batch, and 3) before transaction observers instead of after.
 
 ### Pending
 - Memory management
 - useTransition() compatibility
-- Re-renders from Recoil updates now occur 1) earlier, 2) in sync with React updates in the same batch, and 3) before transaction observers instead of after.
 
 ## 0.5.2 (2021-11-07)